pyproject.toml
[tool.poetry]
name = "tickersentimenttool"
version = "0.1.0"
description = ""
authors = ["nimoes <mhseo2180@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
tweepy = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
